requirements.txt
asynctest==0.13.0
autopep8==1.6.0
black==22.3.0
certifi==2021.10.8
cfgv==3.3.1
charset-normalizer==2.0.12
click==8.1.3
distlib==0.3.4
filelock==3.6.0
identify==2.5.0
idna==3.3
importlib-resources==5.7.1
isort==5.10.1
mypy==0.950
mypy-extensions==0.4.3
nodeenv==1.6.0
numpy==1.22.3
pathspec==0.9.0
platformdirs==2.5.2
pre-commit==2.18.1
promisio==0.1.1
pycodestyle==2.8.0
PyJWT==2.3.0
python-dotenv==0.20.0
PyYAML==6.0
ramda==0.7.4
requests==2.27.1
requests-toolbelt==0.9.1
six==1.16.0
toml==0.10.2
tomli==2.0.1
toolz==0.11.2
typeguard==2.13.3
types-requests==2.27.25
types-urllib3==1.26.14
typing_extensions==4.2.0
urllib3==1.26.9
virtualenv==20.14.1
zipp==3.8.0
